Fixed the bug in which the Today button text changes
The behavior is to go to today whenever someone clicks the button, which
is correct.  However, the text on the button previously changed by
calendar range size, which was incorrect.

I haven't tested this change yet, but it gives every indication of being
a safe change that fixes the bug with a minimum of impact to the system.