Rationalized the code that I changed in the previous commit
Walked through each of the methods in CalendarModel much more carefully
than before, trying to replicate each method while crucially setting the
day of the week to the first day and setting the date to the first of
the month, whenever appropriate.  This should fix the typos I left
during refactoring in the previous commit.

I still haven't tested this code (that is, I haven't even loaded the
webpage produced by this code into a browser), so take it with a
pinch of salt.