Added proper locking around mutable unsafe Calendar objects
The java.util.Calendar class is not thread-safe.  I didn't see any
references to thread-safety in the standard library documentation, but
in a StackOverflow answer (http://stackoverflow.com/a/6245086) Jon Skeet
claims that Calendar is unsafe.  Because the standard library
documentation doesn't say anything about thread safety, but does refer
to fields that aren't fully computed until the object is read, it's
probably safe to assume that Calendar is thread-unsafe and that
synchronization is needed.

As a matter of behavior, before this change switching to day view and
pressing the right arrow usually moved the calendar forward one day but
sometimes moved forward zero or two days (the left arrow did something
similar for moving backwards).  Most often, this kind of odd behavior
occurred when the user double-clicked the button: a double-click was
usually worth two days, but sometimes one day and occasionally four.

Now, if the calendar is on a chunk of time that the user hasn't visited
in this session (the dependency may not be on visiting or not visiting
this session, but on visiting or not visiting ever), a double-click is
more likely than before to be worth only one day.  However, if the user
is in a chunk of time that the user has visited in this session, a
double-click is reliably worth two days.