Fixed bug in which switching to BodyTrack tab only worked partially
Previously, if a user switched to the BodyTrack tab by clicking the
ribbon at the top of the page, the page would load parts of the
BodyTrack app UI but not load the grapher, and not switch the URL to
/app/bodytrack/grapher.  As it turns out, the bug was that, whenever the
renderState function made a recursive call, it forgot to pass along all
but the first parameter, leaving the other parameters undefined.  This
caused an exception on an attempt to set properties on the undefined
params parameter (while inside the recursive call).  The fix was simply
to pass along the same extra parameters as had been passed along before.