Cut down the width of the date axis before channels are added
Previously, the date axis often was too wide for the web browser before
channels were added to it.  Now, the date axis starts out a little too
wide (that is, just wide enough to match the width of the series plot
and the value axis), and then shrinks back to the correct size when
a channel is added, as before.  So the only difference is the
close-to-correct starting width on a date axis.