Now capturing and returning both the exit code and the string response when making calls to the datastore.
Changed executeDataStore() in BodyTrackHelper to return a DataStoreExecutionResult which contains both the String response and the int exitCode.  Changed the uploadToBodyTrack and uploadJsonToBodyTrack methods in BodyTrackHelper to return a BodyTrackUploadResult so that upload methods in BodyTrackController can now return a proper StatusModel depending on the exitCode of the call to the datastore.  Added TODOs for calls to BodyTrackHelper.uploadToBodyTrack() in other classes.