Backing up current state of changes to the time control
I have not found any bugs in this version, which has switched over
to JodaTime and 1-based months, except that the set of dates displayed
on the date picker is completely wrong in week mode.  Only a single day
shows, and that single day typically isn't even part of the current
week.