Timespan now shows up properly when loading /app/calendar/... URLs
When you load a URL directly, it bypasses fetchState() and goes directly
to renderState(); updateTimespanLabel() was not being called. This diff
detects this case and fetches the timespan label from within
renderState().

I also found that simply fetching the label in renderState() wasn't
enough. This renderState() call was happening during require() fetching
of the app templates, resulting in a nasty race condition. The solution
is to Backbone.history.start() only once all templates are loaded, thus
guaranteeing that $('#currentTimespanLabel') can be found.