Fixed the adjustment of weeks and week years for the Sunday start
In CalendarModel.java, previously there was no adjustment on
getWeekYear(), and getWeek() adjusted the week by adding 1.  Both are
wrong around year boundaries: the week year may change between fromDate
and the week that is exposed to the outside world, and the week may be
adjusted beyond the range of valid weeks by simply adding 1.

The fix was to call plusWeeks(1) to get a date that is one week later
than fromDate, resulting in a correct week year and week.