Fixed a bug in the determination of ISO week number in Java
Previously, the week number of a Sunday was one too low because Sunday
is considered part of the previous week under ISO 8601 week date, in
which weeks start on Monday.

Furthermore, added two stopLoading() calls to calendar/App.js: previously,
whenever the user clicked on a date within the currently selected week,
stopLoading() was never called, and the UI appeared to be loading the
date information over a long (never-ending, in fact) time.

Also removed a TODO that is no longer relevant since Candide fixed the
corresponding bug in fluxtream/fluxtream-app/master:422b85def0.