Partially restored the prior timeline dragging/date change behavior
Previously, dragging the timeline caused the calendar to change its date
based on the min time of the timeline, and didn't include momentum.  I
mostly just restored the code that was removed from TimelineTab.js in
commit c1888e76ff, with a few updates to match the new navigation
scheme.  I still have yet to implement code to ensure that navigating
away from and then back to the timeline tab keeps the timeline tab
bounds consistent.