Fixed the time range in the grapher on a fresh page load
In my previous commit, I accidentally set the initial value of
Calendar.timeRange.updated to false, when it should have been true.
Setting this value to true ensures that the time range is set to
the page-wide value whenever the page is refreshed or loaded from a
copy-pasted URL, and not navigated to from inside the web app.