Fixed the drag to add channel feature in the BodyTrack app
The problem was that Grapher.js tried to access the grapher.sourcesMap
property as simply sourcesMap.

Also, changed the CSS rule to set device names to 100% width in the Add
Channels pane so that it didn't affect channel names - in the previous
version of the code, if we commented out the draggable() call and in so
doing disabled dragging to add a channel, the channel name would move to
the left if the user tried to drag it.  Although I don't expect to
remove the drag to add channel feature, this change to bodytrack.css is
a nice way to ensure it goes more smoothly if we ever do so, and to
tighten up CSS rules as general good practice.