Fixed the bug in which the list view was showing week data one day wrong
Previously, the list view was showing week data that started one day too
late and ended one day too late.  For example, if the calendar bar
showed Jul 8 - Jul 14 2012, the data in the list view started at July 9
and ended at July 15.

The problem was that this code wasn't changed over when we moved from
Monday-based to Sunday-based weeks, and so a week was considered to
begin and end one day late.  I switched to using the JodaTime LocalDate,
which is just a more pleasant API than the Java Calendar, and used
TimeUtils.getBeginningOfWeek() to get the week, so that CalendarModel
and CalendarDataStore are consistent.