Fixed back button support
Previously, the back button brought the user to the right URL, but the
date display in the calendar bar stayed the same.  This change doesn't
leave the page marked as having loaded date information, which causes
Calendar.fetchState() to be called when the back button is pressed.  In
turn, this causes the calendar bar to update correctly.